\name{RStructure-package}
\alias{RStructure-package}
\alias{RStructure}
\docType{package}
\title{
Tools to complement Structure
}
\description{
Contains functions for reading and writing files formatted for
Pritchard's Structure program.
}
\details{
}
}
\author{
Ryan Raaum

Maintainer: Ryan Raaum <code@raaum.org>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[genotypeData:genotypeData-package]{genotypeData}}
}
\examples{
}
